First fixes FlowItem, ObjectNode. May not work.
git-svn-id: file:///Users/arjan/backup/gaphor/gaphor/branches/new-canvas@1100 a8418922-720d-0410-834f-a69b97ada669
